llm_samplers

Module configure

Configuring sampler options This module implements configurable samplers based on keys and values and also parsing strings to configure a sampler.

Unless you’re writing your own implementation, you basically only need to worry about ConfigurableSampler::configure. The built in samplers all implement this trait.

Currently the default implementations aren’t as flexible as we might like and it should be noted that values are converted from string to u64 or f64 before being converted to the actual option type.

Modules

configurable_sampler

Since Rust traits don’t allow calling base default methods from a more specific implementation, the ConfigurableSampler trait default methods are implemented in terms of the functions in this submodule.

Structs

SamplerChainBuilder

SamplerMetadata

Structure that defines a sampler’s metadata.

SamplerOptionMetadata

Structure that defines a sampler option.

SamplerOptions

Enums

BuildSamplersError

ConfigureSamplerError

Sampler option handling errors.

SamplerOptionType

Sampler option types.

SamplerOptionValue

Enum that holds the value for a sampler option.

SamplerOptionValueMut

Enum that holds a mutable reference to a sampler option value. It’s only necessary to worry about this when writing your own samplers and implementing option parsing/handling.

SamplerSlot

Traits

BuildableSampler

ConfigurableNumValue

Numeric values that can be used for configuring samplers.

ConfigurableSampler

Configurable samplers implement this trait. “Configurable” means they allow access to their their options by key/type and allow configuration based on descriptions.

HasSamplerMetadata

Configurable samplers will need to implement this trait. It provides metadata for a sampler like its name, description as well as a list of options and their types. It may also provide a way to directly access and manipulate fields in the sampler’s configuration. The built-in samplers all implement this.

Type Aliases

SamplerFactory