llm_samplers::configure

Trait HasSamplerMetadata

pub trait HasSamplerMetadata<UI = usize, F = f32>
where
    UI: ConfigurableNumValue,
    F: ConfigurableNumValue,
{
    // Provided methods
    fn sampler_metadata(&self) -> SamplerMetadata { ... }
    fn sampler_options(&self) -> SamplerOptions<SamplerOptionValue<'_, UI, F>> { ... }
    fn sampler_options_mut(
        &mut self,
    ) -> SamplerOptions<SamplerOptionValueMut<'_, UI, F>> { ... }
}

Configurable samplers will need to implement this trait. It provides metadata for a sampler like its name, description as well as a list of options and their types. It may also provide a way to directly access and manipulate fields in the sampler’s configuration. The built-in samplers all implement this.

Provided Methods

fn sampler_metadata(&self) -> SamplerMetadata

fn sampler_options(&self) -> SamplerOptions<SamplerOptionValue<'_, UI, F>>

fn sampler_options_mut( &mut self, ) -> SamplerOptions<SamplerOptionValueMut<'_, UI, F>>