llm_samplers

Module prelude

Convenient rexports. The simplest way to use the crate is to just throw a use llm_samplers::prelude::*; into your project.

Modules

flat_bias

freq_presence

greedy

locally_typical

min_p

mirostat

rand_distrib

repetition

sequence_repetition

tail_free

temperature

top_a

top_k

top_p

Structs

Logit

An individual logit with some additional metadata for use by the samplers.

Logits

A collection of Logits. You normally will need to build this from the result of evaluating the LLM.

NilSamplerResources

Empty resource structure for use with samplers that don’t require any resources.

SampleFlatBias

Flat bias sampling

SampleFreqPresence

Presence and frequency penalty sampling

SampleGreedy

Greedy sampling

SampleLocallyTypical

Locally typical sampling

SampleMinP

Min-P sampling

SampleMirostat1

Mirostat V1 sampling

SampleMirostat2

Mirostat V2 sampling

SampleRandDistrib

Random distribution sampling

SampleRepetition

Repetition penalty sampling

SampleSeqRepetition

Sequence Repetition

SampleTailFree

Tail free sampling

SampleTemperature

Temperature sampling

SampleTopA

Top-A sampling

SampleTopK

Top-K sampling

SampleTopP

Top-P sampling

SamplerChain

A list of Samplers that can be run in sequence. It implements Sampler so you can build samplers as modular components. A typical use case would be to have several filtering samplers and then a token-picking sampler as the last item to enable calling Sampler::sample_token on the chain.

SimpleSamplerResources

Simple resources that can provide an RNG and/or last tokens to samplers.

Enums

ConfigureSamplerError

Sampler option handling errors.

LogitsError

Logit errors

SamplerError

Sampler errors

Traits

ConfigurableSampler

Configurable samplers implement this trait. “Configurable” means they allow access to their their options by key/type and allow configuration based on descriptions.

HasSamplerResources

Trait for providing resources to samplers.

Sampler

The main sampler trait.

Type Aliases

L

Type for logits.

TID

Type for token IDs.