llm_samplers::prelude

Trait ConfigurableSampler

pub trait ConfigurableSampler<UI = u32, F = f32>: HasSamplerMetadata<UI, F>
where
    UI: ConfigurableNumValue,
    F: ConfigurableNumValue,
{
    // Provided methods
    fn set_option(
        &mut self,
        key: &str,
        val: SamplerOptionValue<'_>,
    ) -> Result<()> { ... }
    fn pre_set_option(
        &mut self,
        md: &SamplerOptionMetadata,
        val: &mut SamplerOptionValue<'_>,
    ) -> Result<()> { ... }
    fn post_set_option(&mut self, md: &SamplerOptionMetadata) -> Result<()> { ... }
    fn get_option(&self, key: &str) -> Result<SamplerOptionValue<'_>> { ... }
    fn configure(&mut self, s: &str) -> Result<()> { ... }
}

Configurable samplers implement this trait. “Configurable” means they allow access to their their options by key/type and allow configuration based on descriptions.

There are default implementations for for all the methods, so in the general case you will only need to implement the OPTIONS constant with the option definitions.

Provided Methods

fn set_option(&mut self, key: &str, val: SamplerOptionValue<'_>) -> Result<()>

Given an option key and SamplerOptionValue attempts to set the option to the specified value.

The default implementation will call Self::pre_set_option before setting the option and Self::post_set_option afterward. For an example of the latter you can look at the Mirostat samplers: when tau is set, they’ll automatically set mu to tau * 2.

fn pre_set_option( &mut self, md: &SamplerOptionMetadata, val: &mut SamplerOptionValue<'_>, ) -> Result<()>

Called before an option is set and is passed a mutable reference to the SamplerOptionValue. It is also passed an index into the options definition list.

The default implementation is a no-op.

fn post_set_option(&mut self, md: &SamplerOptionMetadata) -> Result<()>

Called before an option is set is passed an index into the options definition list.

The default implementation is a no-op.

fn get_option(&self, key: &str) -> Result<SamplerOptionValue<'_>>

Gets an option by name.

fn configure(&mut self, s: &str) -> Result<()>

Updates a sampler’s configurable options based on a string in the format:

key1=value1:key2=value2:keyN=valueN

The key be a prefix of the option name as long as it’s not ambiguous. It’s also possible to just specify the value, which is equivalent to =value (i.e. a blank key name).

Values in this default implementation cannot contain = or : and whitespace at the beginning and end of parts are stripped.