llm_samplers::prelude

Trait Sampler

pub trait Sampler:
    Debug
    + Send
    + Sync {
    // Required method
    fn sample<'a>(
        &mut self,
        res: &mut dyn HasSamplerResources,
        logits: &'a mut Logits,
    ) -> Result<&'a mut Logits>;

    // Provided methods
    fn sampled_token_id(&self) -> Option<TID> { ... }
    fn sample_token(
        &mut self,
        res: &mut dyn HasSamplerResources,
        logits: &mut Logits,
    ) -> Result<Option<TID>> { ... }
}

The main sampler trait.

Required Methods

fn sample<'a>( &mut self, res: &mut dyn HasSamplerResources, logits: &'a mut Logits, ) -> Result<&'a mut Logits>

Runs the Sampler. Depending on the type of Sampler, this may produce a token id.

Provided Methods

fn sampled_token_id(&self) -> Option<TID>

Returns the last sampled token id if available.

A default implemenation is provided which simply returns None.

fn sample_token( &mut self, res: &mut dyn HasSamplerResources, logits: &mut Logits, ) -> Result<Option<TID>>

Run the sampler and return the last sampled token id if available.

A default implementation is provided which just calls Sampler::sample followed by Sampler::sampled_token_id().