llm_samplers::samplers::rand_distrib

Struct SampleRandDistrib

pub struct SampleRandDistrib { /* private fields */ }

Random distribution sampling

A fancy way of saying the sampler selects a token based on the probabilities. For example, if token X has twice the probability value of token Y then we can say token X will be twice as likely to be randomly selected by this sampler.

Properties:

Modifies logits

Selects a token

Parameters:

(none)

Implementations

impl SampleRandDistrib

pub fn new() -> Self