llm_samplers

Module types

Types and traits.

Structs

Logit

An individual logit with some additional metadata for use by the samplers.

Logits

A collection of Logits. You normally will need to build this from the result of evaluating the LLM.

NilSamplerResources

Empty resource structure for use with samplers that don’t require any resources.

SamplerChain

A list of Samplers that can be run in sequence. It implements Sampler so you can build samplers as modular components. A typical use case would be to have several filtering samplers and then a token-picking sampler as the last item to enable calling Sampler::sample_token on the chain.

SimpleSamplerResources

Simple resources that can provide an RNG and/or last tokens to samplers.

Enums

LogitsError

Logit errors

SamplerError

Sampler errors

Traits

HasSamplerResources

Trait for providing resources to samplers.

Sampler

The main sampler trait.

Type Aliases

L

Type for logits.

TID

Type for token IDs.