llm_samplers::types

Trait HasSamplerResources

pub trait HasSamplerResources: Debug {
    // Provided methods
    fn with_rng_mut(
        &mut self,
        _fun: &mut dyn FnMut(&mut dyn RngCore),
    ) -> Result<(), SamplerError> { ... }
    fn with_last_tokens(
        &self,
        _fun: &mut dyn FnMut(&[TID]),
    ) -> Result<(), SamplerError> { ... }
    fn with_last_tokens_mut(
        &mut self,
        _fun: &mut dyn FnMut(&mut Vec<TID>),
    ) -> Result<(), SamplerError> { ... }
}

Trait for providing resources to samplers.

Provided Methods

fn with_rng_mut( &mut self, _fun: &mut dyn FnMut(&mut dyn RngCore), ) -> Result<(), SamplerError>

Allows a sampler to mutably access the RNG (if present).

fn with_last_tokens( &self, _fun: &mut dyn FnMut(&[TID]), ) -> Result<(), SamplerError>

Allows a sampler to immutably access the last tokens (if present).

fn with_last_tokens_mut( &mut self, _fun: &mut dyn FnMut(&mut Vec<TID>), ) -> Result<(), SamplerError>

Allows a sampler to mutably access the last tokens (if present).